+++ eStage/src/components/OffersDashboard.vue
@@ -33,7 +33,7 @@
       <!-- Grid -->
       <div className="mt-6 md:grid md:grid-cols-4 gap-0" v-for="offre in offres" :key="offre._id">
         <div className="flex md:border-b-4 md:p-1 md:border-slate-300">
-          <div className="bg-red-600 w-1 h-12 mr-2 md:mr-6" ></div>
+          <div className="bg-red-600 w-1 h-12 mr-2 md:mr-6"></div>
           <div className="hidden md:block bg-red-600 p-3 w-12 h-12 rounded-lg">
             <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
               <path fill="#ffffff"
@@ -45,8 +45,10 @@
             <h4 className="text-lg">{{ offre.description }}</h4>
           </div>
         </div>
-        <div className="text-black md:border-b-4 md:p-1 md:border-slate-300 text-lg flex items-center">{{ offre.enterprise.city }}</div>
-        <div className="text-black md:border-b-4 md:p-1 md:border-slate-300 text-lg flex items-center">{{ new Date(offre.startDate).toLocaleDateString('fr-CA')
+        <div className="text-black md:border-b-4 md:p-1 md:border-slate-300 text-lg flex items-center">{{
+        offre.enterprise.city }}</div>
+        <div className="text-black md:border-b-4 md:p-1 md:border-slate-300 text-lg flex items-center">{{ new
+        Date(offre.startDate).toLocaleDateString('fr-CA')
           }}</div>
         <div className="text-lg flex justify-between items-center md:border-b-4 md:p-1 md:border-slate-300">
           <button @click="accepterOffre(offre)" className="w-6 md:w-8 lg:hidden xl:hidden">
@@ -116,12 +118,14 @@ const accepterOffre = async (offre) => {
 };
 
 const supprimerOffre = async (id) => {
-  // Appelle la méthode Delete du service en passant l'identifiant de l'offre
-  try {
-    const response = await service.Delete(id);
-    console.log('Offre supprimée avec succès :', response);
-  } catch (error) {
-    console.error(`Erreur lors de la suppression de l'offre :`, error);
+  if (confirm("Desirez-vous vraiment supprimer cela?")) {
+    // Appelle la méthode Delete du service en passant l'identifiant de l'offre
+    try {
+      const response = await service.Delete(id);
+      console.log('Offre supprimée avec succès :', response);
+    } catch (error) {
+      console.error(`Erreur lors de la suppression de l'offre :`, error);
+    }
   }
 };
